claim_extraction/requirements.txt
transformers==4.25.1
spacy==2.3.8
newspaper3k